fastapi-sync/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
jinja2
uvloop